Use proper size_t variables. JIRA: CXX-14376
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@106226 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/util/create_defline.cpp
@@ -2478,8 +2478,8 @@ void CDeflineGenerator::x_SetTitleFromProtein (
                 m_MainTitle.erase (pos + 1);
             }
 
-            int offset = 0;
-            int delta = 0;
+            size_t offset = 0;
+            size_t delta = 0;
             string comma;
             string isoform;
             if (NStr::StartsWith (m_MainTitle, "hypothetical protein")) {
